add new utils file with check_hash and update tests + update workflow (remove update of TEXINPUTS variable)
