+++ tests/test_moodle2amc.py
@@ -19,7 +19,7 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 from amc2moodle.utils.customLogging import customLogger
-from amc2moodle.amc2moodle.test import check_hash
+from amc2moodle.utils.misc import check_hash
 from amc2moodle.moodle2amc import Quiz
 import amc2moodle as amdlpkg
 import os
